requirements.txt
# Python 3.10.0
fastapi==0.111.0
pydantic-settings==2.3.4
python-dotenv==1.0.1
requests==2.32.3
fastapi_cors==0.0.6
pytest==8.2.2